Fix errorelement not working
