+++ src/main.jsx
@@ -1,11 +1,12 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { BrowserRouter } from 'react-router-dom';
+import { BrowserRouter, RouterProvider } from 'react-router-dom';
 import { initReactI18next } from 'react-i18next';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { use } from 'i18next';
 
+import { RefreshToken } from '@/resources/Token/RefreshToken';
 import { Router, RouterPath } from '@/routes/Router';
 
 import { getConfig } from '../config';
@@ -55,9 +56,8 @@ use(initReactI18next).init(
         <Provider store={store}>
           <QueryClientProvider client={queryClient}>
             <LoadLocalizations>
-              <BrowserRouter basename={RouterPath.BasePath}>
-                <Router />
-              </BrowserRouter>
+              <RefreshToken />
+              <RouterProvider router={Router}></RouterProvider>
             </LoadLocalizations>
           </QueryClientProvider>
         </Provider>
