+++ src/routes/Router/Router.tsx
@@ -1,8 +1,5 @@
-/* eslint-disable @typescript-eslint/no-floating-promises */
-/* eslint-disable @typescript-eslint/no-misused-promises */
-import { Route, Routes } from 'react-router-dom';
+import { createBrowserRouter, createRoutesFromElements, Route } from 'react-router-dom';
 import * as React from 'react';
-import axios from 'axios';
 
 import { ErrorPage } from '@/resources/ErrorPage/ErrorPage';
 import { ChooseApiPage } from '@/components/apiDelegation/given/ChooseApiPage';
@@ -25,75 +22,43 @@ export enum RouterPath {
   BasePath = 'accessmanagement/ui',
 }
 
-export const Router = () => {
-  const lastRefreshTokenTimestamp = React.useRef(0);
-  const TEN_MINUTES_IN_MILLISECONDS = 600000;
-
-  async function refreshJwtToken() {
-    const timeNow = Date.now();
-    if (timeNow - lastRefreshTokenTimestamp.current > TEN_MINUTES_IN_MILLISECONDS) {
-      lastRefreshTokenTimestamp.current = timeNow;
-      return await axios
-        // TODO: This may fail in AT if axios doesn't automatically change the base url
-        .get('accessmanagement/api/v1/authentication/refresh')
-        .then((response) => response.data)
-        .catch((error) => {
-          !import.meta.env.DEV && (window.location.pathname = '/');
-          console.error(error);
-        });
-    }
-  }
-
-  React.useEffect(() => {
-    const setUpEventListeners = () => {
-      window.addEventListener('mousemove', refreshJwtToken);
-      window.addEventListener('scroll', refreshJwtToken);
-      window.addEventListener('onfocus', refreshJwtToken);
-      window.addEventListener('keydown', refreshJwtToken);
-    };
-    refreshJwtToken();
-    setUpEventListeners();
-  });
-
-  return (
-    <Routes>
-      createRoutesFromElements(
+export const Router = createBrowserRouter(
+  createRoutesFromElements(
+    <Route
+      path='/'
+      errorElement={<ErrorPage />}
+    >
       <Route
-        path='/'
+        path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiOverview}
+        element={<GivenOverviewPage />}
         errorElement={<ErrorPage />}
-      >
-        <Route
-          path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiOverview}
-          element={<GivenOverviewPage />}
-          errorElement={<ErrorPage />}
-        />
-        <Route
-          path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiChooseOrg}
-          element={<ChooseOrgPage />}
-          errorElement={<ErrorPage />}
-        />
-        <Route
-          path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiChooseApi}
-          element={<ChooseApiPage />}
-          errorElement={<ErrorPage />}
-        />
-        <Route
-          path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiExecuteDelegation}
-          element={<ConfirmationPage />}
-          errorElement={<ErrorPage />}
-        />
-        <Route
-          path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiReceipt}
-          element={<ReceiptPage />}
-          errorElement={<ErrorPage />}
-        />
-        <Route
-          path={RouterPath.ReceivedApiDelegations + '/' + RouterPath.GivenApiOverview}
-          element={<ReceivedOverviewPage />}
-          errorElement={<ErrorPage />}
-        />
-      </Route>
-      , ),
-    </Routes>
-  );
-};
+      />
+      <Route
+        path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiChooseOrg}
+        element={<ChooseOrgPage />}
+        errorElement={<ErrorPage />}
+      />
+      <Route
+        path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiChooseApi}
+        element={<ChooseApiPage />}
+        errorElement={<ErrorPage />}
+      />
+      <Route
+        path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiExecuteDelegation}
+        element={<ConfirmationPage />}
+        errorElement={<ErrorPage />}
+      />
+      <Route
+        path={RouterPath.GivenApiDelegations + '/' + RouterPath.GivenApiReceipt}
+        element={<ReceiptPage />}
+        errorElement={<ErrorPage />}
+      />
+      <Route
+        path={RouterPath.ReceivedApiDelegations + '/' + RouterPath.GivenApiOverview}
+        element={<ReceivedOverviewPage />}
+        errorElement={<ErrorPage />}
+      />
+    </Route>,
+  ),
+  { basename: RouterPath.BasePath },
+);
